move GpuOrcScan
