+++ sql-plugin/src/main/301until310-all/scala/com/nvidia/spark/rapids/shims/v2/GpuOrcScan.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.spark311
+package com.nvidia.spark.rapids.shims.v2
 
 import com.nvidia.spark.rapids.{GpuOrcScanBase, RapidsConf}
 import org.apache.hadoop.conf.Configuration
@@ -41,7 +41,7 @@ case class GpuOrcScan(
     partitionFilters: Seq[Expression],
     dataFilters: Seq[Expression],
     rapidsConf: RapidsConf,
-    queryUsesInputFile: Boolean = true)
+    queryUsesInputFile: Boolean = false)
   extends GpuOrcScanBase(sparkSession, hadoopConf, dataSchema, readDataSchema,
     readPartitionSchema, pushedFilters, rapidsConf, queryUsesInputFile) with FileScan {
